Extract to parent `Modules` class for reuse.
